ingredient_detector/api/requirements.txt
fastapi
uvicorn
torch
opencv-python
pillow
pandas
python-dotenv
groq
